Move `useLoans` relevant functions into the hook
If they are only used inside the hook, it's not necessary to have them in a separate helper function.
